requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
apscheduler==3.10.1
requests==2.31.0
pyyaml==6.0.1
psutil==5.9.5
PyJWT==2.8.0
python-socketio==5.11.1
python-engineio==4.9.0
setuptools<81.0.0  # Add this line to avoid the warning